requirements/api.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
alembic
python-dotenv
httpx